+++ src/main/java/com/kpn/dsh/example/envelope/KeyEnvelopeDeserializer.java
@@ -1,4 +1,4 @@
-package com.kpn.dsh.example;
+package com.kpn.dsh.example.envelope;
 
 import java.util.Map;
 import java.nio.charset.Charset;
